+++ pages/creators.js
@@ -1,5 +1,5 @@
 import CreatorCard from "../components/CreatorCard";
-import SEO from "../components/SEO";
+import SEO from "../components/Shared/SEO";
 import Tabs from "../components/Tabs";
 
 import { getCreators } from "../functions/api";
